STASH

+++ cocos/base/allocator/CCAllocatorStrategyGlobalSmallBlock.h
@@ -126,7 +126,7 @@ public:
     
     virtual ~AllocatorStrategyGlobalSmallBlock()
     {
-        for (int i = 0; i <= kMaxSmallBlockPower; ++i)
+        for (size_t i = 0; i <= kMaxSmallBlockPower; ++i)
             if (_smallBlockAllocators[i])
                 ccAllocatorGlobal.deallocate(_smallBlockAllocators[i]);
